fix(serialization): always use uint64_t for size fields on the wire
Unify container, string, wstring and u16string size encoding/decoding
to std::uint64_t (8 bytes) for consistent cross-platform behaviour
between 32-bit and 64-bit architectures. Update anti-overflow checks
accordingly.

+++ lib/public/StormByte/serializable.cxx
@@ -7,12 +7,12 @@ namespace StormByte {
 	// -------------------------------------------------------------------------
 	template<> STORMBYTE_PUBLIC
 	std::vector<std::byte> Serializable<std::string>::SerializeComplex() const noexcept {
-		std::size_t size = m_data.size();
-		Serializable<std::size_t> size_serial(size);
+		std::uint64_t size = static_cast<std::uint64_t>(m_data.size());
+		Serializable<std::uint64_t> size_serial(size);
 		std::vector<std::byte> buffer = size_serial.Serialize();
-		buffer.reserve(buffer.size() + size * sizeof(char));
+		buffer.reserve(buffer.size() + static_cast<std::size_t>(size) * sizeof(char));
 		const std::byte* data_ptr = reinterpret_cast<const std::byte*>(m_data.data());
-		buffer.insert(buffer.end(), data_ptr, data_ptr + size * sizeof(char));
+		buffer.insert(buffer.end(), data_ptr, data_ptr + static_cast<std::size_t>(size) * sizeof(char));
 		return buffer;
 	}
 
@@ -20,27 +20,27 @@ namespace StormByte {
 	StormByte::Expected<std::string, DeserializeError> Serializable<std::string>::DeserializeComplex(std::span<const std::byte> data) noexcept {
 		std::size_t offset = 0;
 
-		if (offset + sizeof(std::size_t) > data.size())
+		if (offset + sizeof(std::uint64_t) > data.size())
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for string size");
 
-		auto expected_size = Serializable<std::size_t>::Deserialize(data.subspan(offset, sizeof(std::size_t)));
+		auto expected_size = Serializable<std::uint64_t>::Deserialize(data.subspan(offset, sizeof(std::uint64_t)));
 		if (!expected_size)
 			return StormByte::Unexpected(expected_size.error());
 
-		std::size_t size = expected_size.value();
-		offset += sizeof(std::size_t);
+		std::uint64_t size = expected_size.value();
+		offset += sizeof(std::uint64_t);
 
 		// Anti-overflow check
-		if (size > data.size() - offset)
+		if (size > static_cast<std::uint64_t>(data.size() - offset))
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for string content");
 
-		std::string result(reinterpret_cast<const char*>(data.data() + offset), size);
+		std::string result(reinterpret_cast<const char*>(data.data() + offset), static_cast<std::size_t>(size));
 		return result;
 	}
 
 	template<> STORMBYTE_PUBLIC
 	std::size_t Serializable<std::string>::SizeComplex(const std::string& data) noexcept {
-		return sizeof(std::size_t) + data.size();
+		return sizeof(std::uint64_t) + data.size();
 	}
 
 	// -------------------------------------------------------------------------
@@ -48,12 +48,12 @@ namespace StormByte {
 	// -------------------------------------------------------------------------
 	template<> STORMBYTE_PUBLIC
 	std::vector<std::byte> Serializable<std::wstring>::SerializeComplex() const noexcept {
-		std::size_t size = m_data.size();
-		Serializable<std::size_t> size_serial(size);
+		std::uint64_t size = static_cast<std::uint64_t>(m_data.size());
+		Serializable<std::uint64_t> size_serial(size);
 		std::vector<std::byte> buffer = size_serial.Serialize();
-		buffer.reserve(buffer.size() + size * sizeof(wchar_t));
+		buffer.reserve(buffer.size() + static_cast<std::size_t>(size) * sizeof(wchar_t));
 		const std::byte* data_ptr = reinterpret_cast<const std::byte*>(m_data.data());
-		buffer.insert(buffer.end(), data_ptr, data_ptr + size * sizeof(wchar_t));
+		buffer.insert(buffer.end(), data_ptr, data_ptr + static_cast<std::size_t>(size) * sizeof(wchar_t));
 		return buffer;
 	}
 
@@ -61,27 +61,27 @@ namespace StormByte {
 	StormByte::Expected<std::wstring, DeserializeError> Serializable<std::wstring>::DeserializeComplex(std::span<const std::byte> data) noexcept {
 		std::size_t offset = 0;
 
-		if (offset + sizeof(std::size_t) > data.size())
+		if (offset + sizeof(std::uint64_t) > data.size())
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for wstring size");
 
-		auto expected_size = Serializable<std::size_t>::Deserialize(data.subspan(offset, sizeof(std::size_t)));
+		auto expected_size = Serializable<std::uint64_t>::Deserialize(data.subspan(offset, sizeof(std::uint64_t)));
 		if (!expected_size)
 			return StormByte::Unexpected(expected_size.error());
 
-		std::size_t size = expected_size.value();
-		offset += sizeof(std::size_t);
+		std::uint64_t size = expected_size.value();
+		offset += sizeof(std::uint64_t);
 
 		// Anti-overflow: size * sizeof(wchar_t) must fit
-		if (size > (data.size() - offset) / sizeof(wchar_t))
+		if (size > static_cast<std::uint64_t>((data.size() - offset) / sizeof(wchar_t)))
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for wstring content");
 
-		std::wstring result(reinterpret_cast<const wchar_t*>(data.data() + offset), size);
+		std::wstring result(reinterpret_cast<const wchar_t*>(data.data() + offset), static_cast<std::size_t>(size));
 		return result;
 	}
 
 	template<> STORMBYTE_PUBLIC
 	std::size_t Serializable<std::wstring>::SizeComplex(const std::wstring& data) noexcept {
-		return sizeof(std::size_t) + data.size() * sizeof(wchar_t);
+		return sizeof(std::uint64_t) + data.size() * sizeof(wchar_t);
 	}
 
 	// -------------------------------------------------------------------------
@@ -89,12 +89,12 @@ namespace StormByte {
 	// -------------------------------------------------------------------------
 	template<> STORMBYTE_PUBLIC
 	std::vector<std::byte> Serializable<std::u16string>::SerializeComplex() const noexcept {
-		std::size_t size = m_data.size();
-		Serializable<std::size_t> size_serial(size);
+		std::uint64_t size = static_cast<std::uint64_t>(m_data.size());
+		Serializable<std::uint64_t> size_serial(size);
 		std::vector<std::byte> buffer = size_serial.Serialize();
-		buffer.reserve(buffer.size() + size * sizeof(char16_t));
+		buffer.reserve(buffer.size() + static_cast<std::size_t>(size) * sizeof(char16_t));
 		const std::byte* data_ptr = reinterpret_cast<const std::byte*>(m_data.data());
-		buffer.insert(buffer.end(), data_ptr, data_ptr + size * sizeof(char16_t));
+		buffer.insert(buffer.end(), data_ptr, data_ptr + static_cast<std::size_t>(size) * sizeof(char16_t));
 		return buffer;
 	}
 
@@ -102,26 +102,26 @@ namespace StormByte {
 	StormByte::Expected<std::u16string, DeserializeError> Serializable<std::u16string>::DeserializeComplex(std::span<const std::byte> data) noexcept {
 		std::size_t offset = 0;
 
-		if (offset + sizeof(std::size_t) > data.size())
+		if (offset + sizeof(std::uint64_t) > data.size())
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for u16string size");
 
-		auto expected_size = Serializable<std::size_t>::Deserialize(data.subspan(offset, sizeof(std::size_t)));
+		auto expected_size = Serializable<std::uint64_t>::Deserialize(data.subspan(offset, sizeof(std::uint64_t)));
 		if (!expected_size)
 			return StormByte::Unexpected(expected_size.error());
 
-		std::size_t size = expected_size.value();
-		offset += sizeof(std::size_t);
+		std::uint64_t size = expected_size.value();
+		offset += sizeof(std::uint64_t);
 
 		// Anti-overflow: size * sizeof(char16_t) must fit
-		if (size > (data.size() - offset) / sizeof(char16_t))
+		if (size > static_cast<std::uint64_t>((data.size() - offset) / sizeof(char16_t)))
 			return StormByte::Unexpected<DeserializeError>("Insufficient data for u16string content");
 
-		std::u16string result(reinterpret_cast<const char16_t*>(data.data() + offset), size);
+		std::u16string result(reinterpret_cast<const char16_t*>(data.data() + offset), static_cast<std::size_t>(size));
 		return result;
 	}
 
 	template<> STORMBYTE_PUBLIC
 	std::size_t Serializable<std::u16string>::SizeComplex(const std::u16string& data) noexcept {
-		return sizeof(std::size_t) + data.size() * sizeof(char16_t);
+		return sizeof(std::uint64_t) + data.size() * sizeof(char16_t);
 	}
 }
